+++ samples/index.php
@@ -3,7 +3,7 @@ include_once 'Sample_Header.php';
 $requirements = array(
     'php'   => array('PHP 5.3.0', version_compare(phpversion(), '5.3.0', '>=')),
     'xml'   => array('PHP extension XML', extension_loaded('xml')),
-    'temp'  => array('Temp folder "<code>' . sys_get_temp_dir() . '</code>" is writable', is_writable(sys_get_temp_dir())),
+    'temp'  => array('Temp folder "<code>' . Settings::getTempDir() . '</code>" is writable', is_writable(Settings::getTempDir())),
     'zip'   => array('PHP extension ZipArchive (optional)', extension_loaded('zip')),
     'gd'    => array('PHP extension GD (optional)', extension_loaded('gd')),
     'xmlw'  => array('PHP extension XMLWriter (optional)', extension_loaded('xmlwriter')),