+++ tests/PhpWord/Tests/_includes/TestHelperDOCX.php
@@ -19,6 +19,7 @@ namespace PhpOffice\PhpWord\Tests;
 
 use PhpOffice\PhpWord\IOFactory;
 use PhpOffice\PhpWord\PhpWord;
+use PhpOffice\PhpWord\Settings;
 
 /**
  * Test helper class
@@ -41,9 +42,9 @@ class TestHelperDOCX
      */
     public static function getDocument(PhpWord $phpWord, $writerName = 'Word2007')
     {
-        self::$file = tempnam(sys_get_temp_dir(), 'PhpWord');
-        if (!is_dir(sys_get_temp_dir() . '/PhpWord_Unit_Test/')) {
-            mkdir(sys_get_temp_dir() . '/PhpWord_Unit_Test/');
+        self::$file = tempnam(Settings::getTempDir(), 'PhpWord');
+        if (!is_dir(Settings::getTempDir() . '/PhpWord_Unit_Test/')) {
+            mkdir(Settings::getTempDir() . '/PhpWord_Unit_Test/');
         }
 
         $xmlWriter = IOFactory::createWriter($phpWord, $writerName);
@@ -52,11 +53,11 @@ class TestHelperDOCX
         $zip = new \ZipArchive;
         $res = $zip->open(self::$file);
         if ($res === true) {
-            $zip->extractTo(sys_get_temp_dir() . '/PhpWord_Unit_Test/');
+            $zip->extractTo(Settings::getTempDir() . '/PhpWord_Unit_Test/');
             $zip->close();
         }
 
-        return new XmlDocument(sys_get_temp_dir() . '/PhpWord_Unit_Test/');
+        return new XmlDocument(Settings::getTempDir() . '/PhpWord_Unit_Test/');
     }
 
     /**
@@ -67,8 +68,8 @@ class TestHelperDOCX
         if (file_exists(self::$file)) {
             unlink(self::$file);
         }
-        if (is_dir(sys_get_temp_dir() . '/PhpWord_Unit_Test/')) {
-            self::deleteDir(sys_get_temp_dir() . '/PhpWord_Unit_Test/');
+        if (is_dir(Settings::getTempDir() . '/PhpWord_Unit_Test/')) {
+            self::deleteDir(Settings::getTempDir() . '/PhpWord_Unit_Test/');
         }
     }